requirements.txt
PyYAML>=6.0.1
Jinja2>=3.1.2
MarkupSafe>=2.1.3
pytest>=7.4.0
pytest-cov>=4.1.0
